Removing usleep()s before reads.
The kerkey does not offer a way to signal that a read will be
successful. So the previous attempt was to sleep for some time and hope
that the read will be successful.

This patch replaces the old logic, by reading in a loop until the read
succeeds. This improves the response times significantly.

Signed-off-by: Christoph Müllner <christophm30@gmail.com>