Fix handling of too long APDUs.
When sending APDUs with an Le >= 0x8000 the kerkey immediately
(without the need to send the whole APDU) responds with 67 00 (invalid length).

So far we cosidered such a behaviour (not letting us send the whole
APDU) as communication error.

This patch allows the behaviour and prints a debug message in this case.

Signed-off-by: Christoph Müllner <christophm30@gmail.com>