Properly handle write request from device if no more data available.
When receiving the status bytes 0x80,0x00 from the device, then it
expects more data to be sent from the host. In the case that we don't
have any more data, there must be a communication error.
Any attempts to retry will end up in endless loops.
This patch properly detects this case and bails out.

Signed-off-by: Christoph Müllner <christophm30@gmail.com>