Move kernel's gpio.h to ext directory.
The new GPIO API has been introduced in mid 2016.
This leaves all kernel before that time (<4.6) with
a compile problem because the API's header file is missing.

Given that such old kernels are still maintained
(e.g. 4.4-LTS has support till 2022) and the fact that
a recent commit made the complete reset handling optional,
a solution for this problem can be justified.

Therefore this patch merges the original GPIO API header
to the ext directory (similar to the PCSC lite headers)
and switches kerkey.c to use that.
That's sufficient to make things compile with older kernel
headers.

Signed-off-by: Christoph Müllner <christophm30@gmail.com>