Refactoring: Split library into pieces.
The current approach of concentrating the code in kerkey.c
was already ignoring good practices and caused maintenance
issues. This patch addresses the everything-in-one-file
problem by providing the required abstractions for gpio and
I2C modules/drivers.

The result is, that we have now lots of small code files with
independent APIs to interact with each other.

Signed-off-by: Christoph Müllner <christoph.muellner@theobroma-systems.com>