Add option no reset.
While it is currently possibly to avoid reset signaling using GPIO,
it is not possible to prevent resets using protocol commands.

This patch adds the configuration "noreset" to handle those cases where
resets are not desired.

Since no ATR can be gathered in those cases a fixed ATR is forged.